refactor(edge): isolate bootstrap websocket upgrade handling
