+++ crates/edge/src/quic_listener/bootstrap/dispatch.rs
@@ -4,12 +4,10 @@ use std::{
 };
 
 use bytes::Bytes;
-use http::{Request, Response, StatusCode, Uri};
+use http::{Request, Response, StatusCode};
 use http_body_util::combinators::BoxBody;
-use hyper::{body::Incoming, client::conn::http1 as client_http1};
-use hyper_util::rt::TokioIo;
+use hyper::body::Incoming;
 use log::warn;
-use spooky_config::backend_endpoint::BackendScheme;
 use spooky_errors::{ProxyError, classify_upstream_proxy_error};
 use spooky_transport::transport_pool::UpstreamTransportPool;
 
@@ -18,7 +16,7 @@ use crate::{
     quic_listener::{
         QUICListener,
         bootstrap::{
-            BootstrapPreparedRoute, bootstrap_error_response,
+            BootstrapPreparedRoute, bootstrap_error_response, dispatch_bootstrap_websocket,
             request::bootstrap_backend_target_for_prepared,
             request::bootstrap_route_target_for_prepared,
         },
@@ -38,7 +36,7 @@ pub(in crate::quic_listener) struct BootstrapDispatchInput<'a> {
     pub(in crate::quic_listener) alt_svc: &'a str,
 }
 
-fn observe_bootstrap_dispatch_failure(
+pub(in crate::quic_listener) fn observe_bootstrap_dispatch_failure(
     prepared_route: &BootstrapPreparedRoute,
     metrics: &Metrics,
     request_start: Instant,
@@ -85,117 +83,6 @@ fn observe_bootstrap_dispatch_failure(
             "Bootstrap upstream error route={} backend={}: {}",
             prepared_route.upstream_name, prepared_route.backend_addr, proxy_err
         );
-    }
-}
-
-async fn dispatch_bootstrap_websocket(
-    input: BootstrapDispatchInput<'_>,
-) -> Result<Response<Incoming>, Response<BoxBody<Bytes, Infallible>>> {
-    if input.prepared_route.endpoint.scheme() != BackendScheme::Http {
-        return Err(bootstrap_error_response(
-            input.alt_svc,
-            StatusCode::BAD_GATEWAY,
-            b"websocket bootstrap requires http upstream\n",
-        ));
-    }
-
-    let backend_target = input.prepared_route.endpoint.authority().to_string();
-    let upstream_path_uri = match Uri::try_from(input.request_path) {
-        Ok(uri) => uri,
-        Err(_) => {
-            return Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::BAD_GATEWAY,
-                b"bad uri\n",
-            ));
-        }
-    };
-    let (mut parts, body) = input.upstream_req.into_parts();
-    parts.uri = upstream_path_uri;
-    let upstream_req = Request::from_parts(parts, body);
-
-    let stream = match tokio::time::timeout(
-        input.backend_timeout,
-        tokio::net::TcpStream::connect(&backend_target),
-    )
-    .await
-    {
-        Ok(Ok(stream)) => {
-            if let Ok(resolved_addr) = stream.peer_addr() {
-                input.metrics.record_backend_connect(
-                    &backend_target,
-                    input.prepared_route.endpoint.authority_host(),
-                    resolved_addr,
-                );
-            }
-            stream
-        }
-        Ok(Err(err)) => {
-            warn!("Bootstrap WebSocket connect error: {}", err);
-            return Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::BAD_GATEWAY,
-                b"upstream error\n",
-            ));
-        }
-        Err(_) => {
-            return Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::GATEWAY_TIMEOUT,
-                b"upstream timeout\n",
-            ));
-        }
-    };
-
-    let io = TokioIo::new(stream);
-    let (mut sender, conn) = match client_http1::handshake(io).await {
-        Ok(v) => v,
-        Err(err) => {
-            warn!("Bootstrap WebSocket handshake setup failed: {}", err);
-            return Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::BAD_GATEWAY,
-                b"upstream error\n",
-            ));
-        }
-    };
-    tokio::spawn(async move {
-        let _ = conn.with_upgrades().await;
-    });
-
-    match tokio::time::timeout(input.backend_timeout, sender.send_request(upstream_req)).await {
-        Ok(Ok(resp)) => Ok(resp),
-        Ok(Err(err)) => {
-            let proxy_err = ProxyError::Transport(err.to_string());
-            observe_bootstrap_dispatch_failure(
-                input.prepared_route,
-                input.metrics,
-                input.request_start,
-                input.request_id,
-                StatusCode::BAD_GATEWAY,
-                &proxy_err,
-            );
-            Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::BAD_GATEWAY,
-                b"upstream error\n",
-            ))
-        }
-        Err(_) => {
-            observe_bootstrap_dispatch_failure(
-                input.prepared_route,
-                input.metrics,
-                input.request_start,
-                input.request_id,
-                StatusCode::GATEWAY_TIMEOUT,
-                &ProxyError::Timeout,
-            );
-            Err(bootstrap_error_response(
-                input.alt_svc,
-                StatusCode::GATEWAY_TIMEOUT,
-                b"upstream timeout\n",
-            ))
-        }
     }
 }
