+++ crates/edge/src/quic_listener/bootstrap/websocket.rs
@@ -1,0 +1,199 @@
+use std::convert::Infallible;
+
+use bytes::Bytes;
+use http::{Request, Response, StatusCode, Uri, response::Builder as ResponseBuilder};
+use http_body_util::{BodyExt, Full, combinators::BoxBody};
+use hyper::{body::Incoming, client::conn::http1 as client_http1, upgrade, upgrade::OnUpgrade};
+use hyper_util::rt::TokioIo;
+use log::{debug, warn};
+use spooky_config::backend_endpoint::BackendScheme;
+use spooky_errors::ProxyError;
+
+use crate::quic_listener::{
+    bootstrap::{BootstrapDispatchInput, BootstrapPreparedRoute, bootstrap_error_response},
+    protocol::is_websocket_upgrade_request,
+};
+
+use super::dispatch::observe_bootstrap_dispatch_failure;
+
+pub(in crate::quic_listener) struct BootstrapWebsocketFlow {
+    pub(in crate::quic_listener) is_websocket_upgrade: bool,
+    pub(in crate::quic_listener) client_upgrade: Option<OnUpgrade>,
+}
+
+fn boxed_full(body: Bytes) -> BoxBody<Bytes, Infallible> {
+    Full::new(body).map_err(|never| match never {}).boxed()
+}
+
+pub(in crate::quic_listener) fn capture_bootstrap_websocket_flow(
+    req: &mut Request<Incoming>,
+    use_h2: bool,
+) -> BootstrapWebsocketFlow {
+    let is_websocket_upgrade = is_websocket_upgrade_request(req, use_h2);
+    let client_upgrade = if is_websocket_upgrade {
+        Some(upgrade::on(&mut *req))
+    } else {
+        None
+    };
+
+    BootstrapWebsocketFlow {
+        is_websocket_upgrade,
+        client_upgrade,
+    }
+}
+
+pub(in crate::quic_listener) async fn dispatch_bootstrap_websocket(
+    input: BootstrapDispatchInput<'_>,
+) -> Result<Response<Incoming>, Response<BoxBody<Bytes, Infallible>>> {
+    if input.prepared_route.endpoint.scheme() != BackendScheme::Http {
+        return Err(bootstrap_error_response(
+            input.alt_svc,
+            StatusCode::BAD_GATEWAY,
+            b"websocket bootstrap requires http upstream\n",
+        ));
+    }
+
+    let backend_target = input.prepared_route.endpoint.authority().to_string();
+    let upstream_path_uri = match Uri::try_from(input.request_path) {
+        Ok(uri) => uri,
+        Err(_) => {
+            return Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::BAD_GATEWAY,
+                b"bad uri\n",
+            ));
+        }
+    };
+    let (mut parts, body) = input.upstream_req.into_parts();
+    parts.uri = upstream_path_uri;
+    let upstream_req = Request::from_parts(parts, body);
+
+    let stream = match tokio::time::timeout(
+        input.backend_timeout,
+        tokio::net::TcpStream::connect(&backend_target),
+    )
+    .await
+    {
+        Ok(Ok(stream)) => {
+            if let Ok(resolved_addr) = stream.peer_addr() {
+                input.metrics.record_backend_connect(
+                    &backend_target,
+                    input.prepared_route.endpoint.authority_host(),
+                    resolved_addr,
+                );
+            }
+            stream
+        }
+        Ok(Err(err)) => {
+            warn!("Bootstrap WebSocket connect error: {}", err);
+            return Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::BAD_GATEWAY,
+                b"upstream error\n",
+            ));
+        }
+        Err(_) => {
+            return Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::GATEWAY_TIMEOUT,
+                b"upstream timeout\n",
+            ));
+        }
+    };
+
+    let io = TokioIo::new(stream);
+    let (mut sender, conn) = match client_http1::handshake(io).await {
+        Ok(v) => v,
+        Err(err) => {
+            warn!("Bootstrap WebSocket handshake setup failed: {}", err);
+            return Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::BAD_GATEWAY,
+                b"upstream error\n",
+            ));
+        }
+    };
+    tokio::spawn(async move {
+        let _ = conn.with_upgrades().await;
+    });
+
+    match tokio::time::timeout(input.backend_timeout, sender.send_request(upstream_req)).await {
+        Ok(Ok(resp)) => Ok(resp),
+        Ok(Err(err)) => {
+            let proxy_err = ProxyError::Transport(err.to_string());
+            observe_bootstrap_dispatch_failure(
+                input.prepared_route,
+                input.metrics,
+                input.request_start,
+                input.request_id,
+                StatusCode::BAD_GATEWAY,
+                &proxy_err,
+            );
+            Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::BAD_GATEWAY,
+                b"upstream error\n",
+            ))
+        }
+        Err(_) => {
+            observe_bootstrap_dispatch_failure(
+                input.prepared_route,
+                input.metrics,
+                input.request_start,
+                input.request_id,
+                StatusCode::GATEWAY_TIMEOUT,
+                &ProxyError::Timeout,
+            );
+            Err(bootstrap_error_response(
+                input.alt_svc,
+                StatusCode::GATEWAY_TIMEOUT,
+                b"upstream timeout\n",
+            ))
+        }
+    }
+}
+
+pub(in crate::quic_listener) fn write_bootstrap_websocket_upgrade(
+    resp_builder: ResponseBuilder,
+    upstream_resp: &mut Response<Incoming>,
+    prepared_route: &BootstrapPreparedRoute,
+    request_start: std::time::Instant,
+    alt_svc: &str,
+    client_upgrade: Option<OnUpgrade>,
+    status: StatusCode,
+) -> Result<Response<BoxBody<Bytes, Infallible>>, hyper::Error> {
+    let client_upgrade = match client_upgrade {
+        Some(upgrade) => upgrade,
+        None => {
+            return Ok(Response::builder()
+                .status(StatusCode::BAD_GATEWAY)
+                .header("alt-svc", alt_svc)
+                .body(boxed_full(Bytes::from_static(b"upgrade setup error\n")))
+                .unwrap_or_else(|_| Response::new(boxed_full(Bytes::from_static(b"error\n")))));
+        }
+    };
+    let upstream_upgrade = upgrade::on(upstream_resp);
+    tokio::spawn(async move {
+        let (client, upstream) = match tokio::try_join!(client_upgrade, upstream_upgrade) {
+            Ok(v) => v,
+            Err(err) => {
+                debug!("Bootstrap WebSocket upgrade join failed: {}", err);
+                return;
+            }
+        };
+        let mut client = TokioIo::new(client);
+        let mut upstream = TokioIo::new(upstream);
+        let _ = tokio::io::copy_bidirectional(&mut client, &mut upstream).await;
+    });
+    crate::runtime::connection::outcome::finish_backend_request_accounting(
+        crate::runtime::connection::outcome::BackendRequestFinishInput {
+            upstream_pool: Some(&prepared_route.upstream_pool),
+            backend_index: Some(prepared_route.backend_index),
+            elapsed: request_start.elapsed(),
+            status: Some(status.as_u16()),
+        },
+    );
+    Ok(resp_builder
+        .body(boxed_full(Bytes::new()))
+        .unwrap_or_else(|_| Response::new(boxed_full(Bytes::new()))))
+}
